added util.js , global.js and unfinished tests files
