+++ src/Globals.js
@@ -1,0 +1,6 @@
+
+
+export const apiV2Url = 'https://nmpi.hbpneuromorphic.eu/api/v2/'
+export const validationV2Url = "https://validation-v2.brainsimulation.eu";
+
+
